Revert "Alerting: add template funcs (#38404)" (#39258)
This reverts commit d6fb0181fbb9033a766d5c54c881c2007a487085.
