+++ pkg/infra/process/root_check.go
@@ -1,4 +1,3 @@
-//go:build !windows
 // +build !windows
 
 package process
